Fix TLS 1.2 regression when openssl >= 1.1.0
* symptom: TLS 1.2 works (e.g., connection string containing "/tls1_2")
with openssl < 1.1.0, but not with openssl >= 1.1.0.  However, running
alpine with -supported argument shows TLSv1.2 support, and the generated
docs show "/tls1_2" support.

* diagnosis: SSL_SUPPORTS_TLSV1_2 is getting defined by configure, but
c-client TLSV1_2 is not.  Alpine thinks TLSv1.2 support is present but
c-client code rejects connection strings containing "/tls1_2" (and
"/tls1_1").  Changing connection strings to "/tls" works, and results in
the use of TLS_client_method() in openssl 1.1.0 which will negotiate the
highest version supported by client and server.  However, the use of
"/tls1_2" indicates that the Alpine user requires TLS 1.2 and "/tls"
doesn't do this.

* fix: Bug appears to have been introduced in be296fed when support for
openssl >= 1.1.0 was added.  The diff for this commit shows that
c-client TLSV1_2 is only defined when running on an "old" (< 1.1.0)
version of openssl.  Add definition of TLSV1_2 to new openssl section in
configure.ac.  Change made to configure.ac, configure regenerated with
autoconf.